1.use range only [min, max] enum id's to improve performance 2.fix dcopy microdata db write error 3.fix dbcopy error if microdata run values are empty (zero row count).

+++ ompp/db/writeMicrodata.go
@@ -458,7 +458,7 @@ func putInsertMicroFrom(
 		}
 
 		// set sql statement microdata values: entity key, attributes value
-		row[0] = cell.Key
+		row[0] = int64(cell.Key)
 
 		for k, av := range cell.Attr {
